-Add auth service which contains all auth fumctions with its implementation. -Add function which create the JWT token. -Add funciton to register new user. -Add fucntion to fetch current user token.

+++ APIsJWTAuthentication/Helpers/JWT.cs
@@ -5,6 +5,6 @@
         public required string? Key { get; set; }
         public required string Issuer { get; set; }
         public required string Audience { get; set; }
-        public int DurationInDays { get; set; }
+        public double DurationInDays { get; set; }
     }
 }
